+++ cmd/template/cluster/provider/capo.go
@@ -72,9 +72,16 @@ func templateClusterOpenstack(ctx context.Context, k8sClient k8sclient.Interface
 					Replicas:      config.OpenStack.WorkerReplicas,
 				},
 			},
-			OIDC: &openstack.OIDC{
-				Enabled: config.OpenStack.EnableOIDC,
-			},
+		}
+
+		if config.OIDC.IssuerURL != "" {
+			flagValues.OIDC = &openstack.OIDC{
+				IssuerURL:     config.OIDC.IssuerURL,
+				CAFile:        config.OIDC.CAFile,
+				ClientID:      config.OIDC.ClientID,
+				UsernameClaim: config.OIDC.UsernameClaim,
+				GroupsClaim:   config.OIDC.GroupsClaim,
+			}
 		}
 
 		configData, err := openstack.GenerateClusterValues(flagValues)
@@ -143,9 +150,6 @@ func templateDefaultAppsOpenstack(ctx context.Context, k8sClient k8sclient.Inter
 		flagValues := openstack.DefaultAppsConfig{
 			ClusterName:  config.Name,
 			Organization: config.Organization,
-			OIDC: &openstack.OIDC{
-				Enabled: config.OpenStack.EnableOIDC,
-			},
 		}
 
 		configData, err := openstack.GenerateDefaultAppsValues(flagValues)